requirements.txt
fastapi 
uvicorn 
pydantic 
tensorflow 
pandas 
numpy 
fastapi[all] 
pillow
python-dotenv
